snapshot 최적화

+++ app/src/main/java/com/swu/dimiz/ogg/contents/listset/ListsetViewModel.kt
@@ -773,7 +773,7 @@ class ListsetViewModel(private val repository: OggRepository) : ViewModel() {
                 .addOnSuccessListener { Timber.i("DocumentSnapshot successfully updated!") }
                 .addOnFailureListener { e -> Timber.i("Error updating document", e) }
 
-            washingtonRef.update("aim", _aimCo2.value?.toDouble())
+            washingtonRef.update("aim", _aimCo2.value?.toFloat())
                 .addOnSuccessListener { Timber.i("DocumentSnapshot successfully updated!") }
                 .addOnFailureListener { e -> Timber.i("Error updating document", e) }
